fix(manager): remove duplicate audio review header
