+++ apps/manager/.env.ci
@@ -1,0 +1,8 @@
+# CI placeholder values. Production and local secrets come from Railway/Doppler.
+MUX_TOKEN_ID=ci-placeholder
+MUX_TOKEN_SECRET=ci-placeholder
+OPENROUTER_API_KEY=ci-placeholder
+STRAPI_URL=http://localhost:1337
+STRAPI_API_TOKEN=ci-placeholder
+STRAPI_INTERNAL_API_TOKEN=ci-placeholder
+MANAGER_API_KEY=ci-placeholder
