+++ apps/manager/src/app/globals.css
@@ -3755,71 +3755,6 @@ body.jobs-standalone .jobs-step-job-id {
   flex: 0 0 auto;
 }
 
-.jobs-audio-review {
-  margin-top: 16px;
-  padding-top: 16px;
-  border-top: 1px solid rgba(95, 87, 76, 0.14);
-}
-
-.jobs-audio-review-header {
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-  gap: 12px;
-  flex-wrap: wrap;
-}
-
-.jobs-audio-review-title {
-  margin: 0;
-  color: #2a241b;
-  font-size: 0.96rem;
-  font-weight: 700;
-  line-height: 1.2;
-}
-
-.jobs-audio-review-grid {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  gap: 12px 20px;
-  margin-top: 10px;
-}
-
-.jobs-audio-review-item {
-  display: grid;
-  gap: 6px;
-  min-width: 0;
-}
-
-.jobs-audio-review-link {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  width: fit-content;
-  min-height: 32px;
-  padding: 0 12px;
-  border-radius: 8px;
-  border: 1px solid #d8cdbd;
-  background: #f3ece1;
-  color: #4f483e;
-  font-size: 0.82rem;
-  font-weight: 700;
-  text-decoration: none;
-  transition:
-    transform 120ms ease,
-    box-shadow 120ms ease,
-    border-color 120ms ease;
-}
-
-.jobs-audio-review-link:hover,
-.jobs-audio-review-link:focus-visible {
-  transform: translateY(-1px);
-  border-color: #c4b6a2;
-  box-shadow: 0 6px 12px rgba(12, 17, 29, 0.14);
-  color: #2f2a23;
-  text-decoration: none;
-  outline: none;
-}
-
 .jobs-inline-icon-button {
   display: inline-flex;
   align-items: center;
@@ -5263,10 +5198,6 @@ body.jobs-standalone .jobs-step-job-id {
   }
 
   .jobs-embedding-sync-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .jobs-audio-review-grid {
     grid-template-columns: 1fr;
   }
