+++ apps/manager/src/features/jobs/live-job-detail-header.tsx
@@ -22,7 +22,6 @@ import {
   getDisplayedJobStatus,
   getLanguageBadges,
 } from "@/features/jobs/jobs-table-presenter"
-import { AudioReviewLinks } from "./audio-review-links"
 
 type LiveJobDetailHeaderProps = {
   job: JobRecord
@@ -269,7 +268,6 @@ export function LiveJobDetailHeader({
           </div>
         </div>
       </div>
-      <AudioReviewLinks job={job} />
     </section>
   )
 }
